Exercicio 8: criando varios pipes e movendo todos eles.

+++ src/com/example/cursojumpbird/ScreenHelper.java
@@ -20,7 +20,7 @@ public class ScreenHelper {
 		return metrics.heightPixels;
 	}
 
-	public float getWidth() {
+	public int getWidth() {
 		return metrics.widthPixels;
 	}
